+++ packages/riverpod_helper/lib/src/prefs_alive_riverpod/prefs_double_pod/prefs_double_pod.g.dart
@@ -51,7 +51,7 @@ final class PrefsAliveDoublePodProvider
 }
 
 String _$prefsAliveDoublePodHash() =>
-    r'efb26efdc4218ccccc8e8cd5a00684a5a08c5bda';
+    r'7d2259168fb53efd7e2c6ea2f651b9c52b5d85cc';
 
 final class PrefsAliveDoublePodFamily extends $Family
     with