+++ packages/riverpod_helper/lib/src/prefs_alive_riverpod/prefs_int_pod/prefs_int_pod.g.dart
@@ -50,7 +50,7 @@ final class PrefsAliveIntPodProvider
   }
 }
 
-String _$prefsAliveIntPodHash() => r'd586abb5ee4b2a341c61d8dcdb0781f2fa03595d';
+String _$prefsAliveIntPodHash() => r'f1ffb166dc0225336ee2c92514d4fefd2cdf0ade';
 
 final class PrefsAliveIntPodFamily extends $Family
     with