+++ packages/riverpod_helper/lib/src/prefs_alive_riverpod/prefs_str_lst_pod/prefs_str_lst_pod.g.dart
@@ -51,7 +51,7 @@ final class PrefsAliveStrLstPodProvider
 }
 
 String _$prefsAliveStrLstPodHash() =>
-    r'7b4e67a4f76975b88b132debd0614a0e3df7937c';
+    r'23fee276ebf39ff8dbbdb18b03e3034b55f0a141';
 
 final class PrefsAliveStrLstPodFamily extends $Family
     with