+++ packages/riverpod_helper/lib/src/prefs_alive_riverpod/prefs_string_pod/prefs_string_pod.g.dart
@@ -51,7 +51,7 @@ final class PrefsAliveStringPodProvider
 }
 
 String _$prefsAliveStringPodHash() =>
-    r'5837b57358c0d5c87e6bd95a7fa9c1efb5719a96';
+    r'd4e3bbdc20ceca82e8bdef8c29197a3955355825';
 
 final class PrefsAliveStringPodFamily extends $Family
     with