+++ packages/riverpod_helper/lib/src/prefs_riverpod/prefs_int_pod/prefs_int_pod.g.dart
@@ -50,7 +50,7 @@ final class PrefsIntPodProvider
   }
 }
 
-String _$prefsIntPodHash() => r'afbe5c95a48d660d2a1f7821ddee769e0a98a03e';
+String _$prefsIntPodHash() => r'5f93f61505e09ac389c4c39e15c9f1e3a60e89cf';
 
 final class PrefsIntPodFamily extends $Family
     with