+++ packages/riverpod_helper/lib/src/prefs_riverpod/prefs_str_lst_pod/prefs_str_lst_pod.g.dart
@@ -50,7 +50,7 @@ final class PrefsStrLstPodProvider
   }
 }
 
-String _$prefsStrLstPodHash() => r'fd2878ce4cd2ffebd7ba64941b49f07fd4263442';
+String _$prefsStrLstPodHash() => r'2a6de4d4cf374d5593ce626eb7b474f0c8377f5d';
 
 final class PrefsStrLstPodFamily extends $Family
     with