+++ packages/riverpod_helper/lib/src/prefs_riverpod/prefs_string_pod/prefs_string_pod.g.dart
@@ -50,7 +50,7 @@ final class PrefsStringPodProvider
   }
 }
 
-String _$prefsStringPodHash() => r'f6f85f04947d418e4b42d2769b0c9654d89a9bd2';
+String _$prefsStringPodHash() => r'e0db95f163b0874fc4c4fd4a46381345c91e2bb6';
 
 final class PrefsStringPodFamily extends $Family
     with